🐞 fix: 修复小错误

+++ sample/edm_sampler.py
@@ -51,4 +51,4 @@ def edm_sampler(
             d_prime = (x_next - denoised) / t_next
             x_next = x_hat + (t_next - t_hat) * (0.5 * d_cur + 0.5 * d_prime)
             
-    return x_next
+    return x_next.to(torch.float32)
